pyproject.toml
[project]
name = "gpt-researcher"
version = "0.8.5"
description = "GPT Researcher is an autonomous agent designed for comprehensive online research on a variety of tasks."
authors = [{ name = "Assaf Elovic", email = "assaf.elovic@gmail.com" }]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = "MIT"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "colorama>=0.4.6",
    "duckduckgo_search>=4.1.1",
    "md2pdf>=1.0.1",
    "openai>=1.3.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "uvicorn>=0.24.0.post1",
    "pydantic>=2.5.1",
    "fastapi>=0.104.1",
    "python-multipart>=0.0.6",
    "markdown>=3.5.1",
    "langchain>=0.3.18,<0.4",
    "langgraph>=0.2.73,<0.3",
    "langchain_community>=0.3.17,<0.4",
    "langchain-openai>=0.3.6,<0.4",
    "tavily-python>=0.2.8",
    "permchain>=0.0.6",
    "arxiv>=2.0.0",
    "PyMuPDF>=1.23.6",
    "requests>=2.31.0",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "SQLAlchemy>=2.0.28",
    "mistune>=3.0.2,<4",
    "htmldocx>=0.0.6,<0.0.7",
    "python-docx>=1.1.0,<2",
    "lxml[html_clean]>=4.9.2",
    "unstructured>=0.13,<0.16",
    "tiktoken>=0.7.0",
    "json-repair>=0.29.8,<0.30",
    "json5>=0.9.25,<0.10",
    "loguru>=0.7.2,<0.8",
    "websockets~=13.1",
    "langgraph-cli[inmem]>=0.1.71",
    "azure-storage-blob>=12.24.1",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "-v"
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
